backend/requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
anthropic
python-multipart
pydantic
python-dotenv
